+++ include/votca/xtp/atom.h
@@ -24,16 +24,14 @@
 #define VOTCA_XTP_ATOM_H
 
 // Standard includes
+#include "eigen.h"
 #include <exception>
 #include <map>
 #include <string>
-
-// Local VOTCA includes
-#include "checkpointreader.h"
-#include "checkpointwriter.h"
-
+#include <votca/tools/types.h>
 namespace votca {
 namespace xtp {
+class CptTable;
 class Atom {
  public:
   struct data {
@@ -78,7 +76,7 @@ class Atom {
     return out;
   }
 
-  void SetupCptTable(CptTable& table) const;
+  static void SetupCptTable(CptTable& table);
 
   void WriteData(data& d) const;