+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Version 2022-dev
 ================
 
 -  Updated coordinate precision (#638)
+-  Made SetupCptTable static (#650)
 -  using libecpint for pseudo potentials (#588)
 
 Version 2021-rc.2 (released XX.01.21)